mailroom-pt3.py has one comprehension now, have a couple more candidate spots on deck
