+++ students/nick_miller/noodling/janchipchase.txt
@@ -1,0 +1,9 @@
+
+Dearest Jan,
+
+Thank you for your generous support!
+We appreciate your donation(s), which total $1504.10 to date!
+
+Sincerest regards,
+
+The Foundation
