+++ students/nick_miller/noodling/jeffstaple.txt
@@ -1,0 +1,9 @@
+
+Dearest Jeff,
+
+Thank you for your generous support!
+We appreciate your donation(s), which total $40.00 to date!
+
+Sincerest regards,
+
+The Foundation
