+++ students/nick_miller/noodling/sketchpad3.py
