+++ students/nick_miller/noodling/takashimurakami.txt
@@ -1,0 +1,9 @@
+
+Dearest Takashi,
+
+Thank you for your generous support!
+We appreciate your donation(s), which total $10.50 to date!
+
+Sincerest regards,
+
+The Foundation
