+++ students/nick_miller/noodling/virgilabloh.txt
@@ -1,0 +1,9 @@
+
+Dearest Virgil,
+
+Thank you for your generous support!
+We appreciate your donation(s), which total $345.68 to date!
+
+Sincerest regards,
+
+The Foundation
